requirements.txt
# Pip requirements for Google Cloud Functions.
# Must match install_requires from setup.py.
google-cloud-bigquery
google-cloud-pubsub<2.0.0dev
google-cloud-firestore
google-cloud-secret-manager<2.0.0dev
jsonschema[format]
requests
pyyaml
jinja2
python-dateutil
jq@git+https://github.com/spbnick/jq.py.git@1.1.2.post1
kcidb-io@git+https://github.com/kernelci/kcidb-io.git@8190c1f5bee4140bc91cd1432191ca8d2acb9107
